Addition of Enum class to check user's answer
Finishing print question method in the fill in the blank class and addition of enum class to tell whether user is correct or incorrect or tell the user that they have chose a question not within the range of multi-choice question. Also  in the Question class there will be another abstract method called checkUserAnswer.