Implement conversion to special German character
Since the requirement states that users dont have to reconfig their
computer to run the programm. This will be a drawback to type German.
The solution will be let user type in symbol then converted it to
special German character. The commit also cleans up some part that is
not neccessary. There is an addion of guide to print special character
method in add to remind user how to type special German character