Fixes bug displaying two of the same answers
Fixed bug where two of possible answers were the same. This was caused because if the word that was randomly grabbed from the word list was the same as the answer than it would not check and add it. Added a check for this and and reiterated the loop if it does happen.