Addition of new mode called learning mode
The new class will be used for learning mode as opposed to testing mode.
Some of the import were removed because it is redundant. Deletion some
comment so it would not confuse the reader. There is a static field
added into the main class because it will be better for other class to
access available word without passing the parameter around.