Change Learning mode class structure
This changes will include code clean up, change data type of some field.
Some comments are removed since it is not explained properly. Refactor
the name of some variable so it will be descriptive. Remove random
object since we are using shuffle method from collections because it is
better for performance