Addition of User, UserList.txt and UserListGenerator
Addition of the User class. This class keeps track of users name and the index where they were up to in the learning mode. Also added the UserList.txt file which keeps a list of the users that have used this program, this file is both read in and written to. Finally also added the UserListGenerator class which similar to the AvailableWord class handles all the file input and output and generates the user list from the UserList.txt file.