Fixes Bug that caused An IndexOoutofBounds  Exception
Fixed a bug that cause an IndexOutofBound exception which was caused  because there was a method call to in the printWords() method when no words had been added. as the generateQuestions was called after the printWords() method. Also tidied up the code a little, removing extra whitespace and keeping the scope brackets consistent.  According to the Google Java Standard Style Guide https://google.github.io/styleguide/javaguide.html