Added a new Word List in LearningMode
Added a new word list variable as the word list in the vocab learning is shuffled. Refactored the code to use the new word list variable. Also changed the generateQuestion() method to use the lastIndex(an int value that represent the word index that the user has learned upto) to determine the question for the the startSmallTest() method