Fixes IndexOutofBounds exception in getUserNumberWordsLearn() method
Fixed IndexOutofBounds exception. This was cause as the user was able to enter a number if they had already had saved user in the UserList.txt then they could enter a number when prompted on how much words they wanted to learn. Fixed this by subtracting the user's last index variable from the numberWordsAvailable variable. Also changed the getCurrentUser() method from LearningMode to getCurrentUserLastIndex() as we don't need to return a user object just the index.