Added score and fixed bug in learning mode
Added a score so user known how much of the words they got right at the end of the learning mode. Also fixed a bug where the special German characters would not convert when the user was entering in the learning mode. This bug was because the method convertSpecialChar() was not being called.